allow float as send/receive and label names in iemguis
closes https://github.com/pure-data/pure-data/issues/1081

+++ src/g_all_guis.c
@@ -200,7 +200,7 @@ t_symbol *iemgui_new_dogetname(t_iemgui *iemgui, int indx, t_atom *argv)
     else if (IS_A_FLOAT(argv, indx))
     {
         char str[80];
-        sprintf(str, "%d", (int)atom_getfloatarg(indx, 100000, argv));
+        sprintf(str, "%g", atom_getfloatarg(indx, 100000, argv));
         return (gensym(str));
     }
     else return (gensym("empty"));
@@ -638,21 +638,21 @@ int iemgui_dialog(t_iemgui *iemgui, t_symbol **srl, int argc, t_atom *argv)
         srl[0] = atom_getsymbolarg(7, argc, argv);
     else if(IS_A_FLOAT(argv,7))
     {
-        sprintf(str, "%d", (int)atom_getfloatarg(7, argc, argv));
+        sprintf(str, "%g", atom_getfloatarg(7, argc, argv));
         srl[0] = gensym(str);
     }
     if(IS_A_SYMBOL(argv,8))
         srl[1] = atom_getsymbolarg(8, argc, argv);
     else if(IS_A_FLOAT(argv,8))
     {
-        sprintf(str, "%d", (int)atom_getfloatarg(8, argc, argv));
+        sprintf(str, "%g", atom_getfloatarg(8, argc, argv));
         srl[1] = gensym(str);
     }
     if(IS_A_SYMBOL(argv,9))
         srl[2] = atom_getsymbolarg(9, argc, argv);
     else if(IS_A_FLOAT(argv,9))
     {
-        sprintf(str, "%d", (int)atom_getfloatarg(9, argc, argv));
+        sprintf(str, "%g", atom_getfloatarg(9, argc, argv));
         srl[2] = gensym(str);
     }
     if(init != 0) init = 1;
